Trac #8664: #8664 Make ext. modules that are linked against MPIR depend on its header(s).
Since there is a huge number of modules that are linked against libgmp.so.*
(and a slightly smaller that are linked against libgmpxx.so), but currently
don't explicitly depend on GMP/MPIR, add these dependencies in a generic way.

Otherwise on an upgrade (or when just installing the new spkg and running
'sage -b' rather than 'sage -ba'), those modules won't get rebuilt and would
still have the old shared libraries recorded, which leads to import errors
or all kinds of failures (segmentation faults etc.).

Only addresses the MPIR upgrade; other spkgs' libraries and their headers
could be added later, but I'm yet unsure if we should do that for *all*
libraries (which is safe and easy, but perhaps inefficient).

--

Still waiting for Cython pragmas ("clib" etc.) that would allow us to imple-
ment this in a more modular but also even more generic way...

--

Rebased on Sage 4.7.2.alpha2 / #11376.